--- pkg/goanalysis/runner_facts.go
@@ -1,16 +1,0 @@
-package goanalysis
-
-import (
-	"go/types"
-	"reflect"
-)
-
-type objectFactKey struct {
-	obj types.Object
-	typ reflect.Type
-}
-
-type packageFactKey struct {
-	pkg *types.Package
-	typ reflect.Type
-}
